refactor: BasketRepository 통합
FDORA-152
